Refactor AOT classes and move to drop-in JsAot
